backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.2
pydantic==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
beautifulsoup4==4.12.3
requests==2.32.3
spacy==3.7.5
sentence-transformers==3.0.1
scikit-learn==1.5.0
transformers==4.41.2
torch==2.3.1
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
bcrypt==4.1.3
